requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dependency-injector==4.45.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.7
fonttools==4.55.8
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
jiter==0.8.2
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
msal==1.31.1
msal-extensions==1.2.0
mysqlclient==2.2.7
numpy==2.2.2
openai==1.61.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
py-ulid==1.0.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
azure-ai-textanalytics==5.3.0
fastapi-utilities==0.3.0
firebase_admin==6.6.0
faiss-cpu==1.10.0
pyaudio==0.2.14
audioop-lts==0.2.1
SpeechRecognition==3.14.1